refactor: extract HardwareCompatibilityMatrix + useOliveStream, update audit doc
- Extracted HardwareCompatibilityMatrix (359 lines, 7 props) from
  IHVIntegrationPanel.tsx — interactive heatmap table + footer legend.
  IHV panel reduced from 1592 to 1282 lines (−310, 19%). (#121)
- Extracted useOliveStream hook (398 lines) from ExecutionWorkspace.tsx —
  full SSE lifecycle: connectSSE with log/metrics/done listeners,
  exponential backoff reconnect (10 att / 30s max), graceful cancel
  handling, job history persistence. Component reduced from 1515 to
  1174 lines (−341, 22%). Cumulative reduction with prior OwrExportOverlay
  extraction: 1753 → 1174 (−579, 33%). (#120)
- Removed 11 unused imports, 7 state/ref declarations across both files.
- Exported OptimizationPassValidation interface from IHV panel.
- Closed #139 (PluginInstallBlock verified DRY — reused 6×).
- Updated docs/Tech Debt & Issues.md: 24 closed, 30 open, all 4
  complexity extractions complete with final line counts.
- Full CI validated: tsc clean, ESLint zero warnings, 1773/1774 tests pass.
